+++ libs/cardano-ledger-core/CHANGELOG.md
@@ -3,7 +3,11 @@
 ## 1.0.1.0
 
 * Add `ToJSON`/`FromJSON` instances for `CompactForm Coin`, `SafeHash` and `TxId`
-* Add `ToJSON` instances for `Ptr`, `CertIx`, `TxIx`, `Trip` and `UMap`
+* Add `ToJSON` instances for:
+  * `Ptr`, `CertIx`, `TxIx`
+  * `Trip` and `UMap`
+  * `DeltaCoin`
+  * `InstantaneousRewards`, `FutureGenDeleg`, `PState`, `DState` and `DPState`.
 
 ## 1.0.0.0